Rettelser på diverse classes

+++ SWC/Passiv.java
@@ -3,9 +3,11 @@ import java.io.*;
 
 public class Passiv extends Kontingent  {
    
-   public Passiv(double pris){ super((int)pris); }   
+   // denne skal nok bare implementeres direkte i kontigent klassen
+   
+   public Passiv(double pris){ super(pris); }   
    
    //polymorphisme
    //type casting to int because it might be lossy
-   public double getPris() { return (int)super.getPris() + 500.0; } 
+   public double getPris() { return super.getPris() + 500.0; } 
 }
